requirements.txt
apache-airflow

# testing
pytest

# dev
jupyter
notebook
